fix popup is unexpectedly closed on CursorHold
CursorHold event triggered CursorMoved event. And the CursorMoved event
closed the window unexpectedly. This happens only on Vim because it's
related to a preview window. When a preview window is open, the global
cursor position may change. In the case, checking if cursor position
moved or not does not work properly since the popup window remembers the
position before opening the preview window.