Add mapping to yank current commit hash
As you can't enter a Vim popup window, you can't easily yank the current
commit hash, and that seems a natural thing to want to do (I certainly
found myself wanting to do it soon after starting to use git-messenger).

To address this, map 'c' to yank the current commit hash to v:register
(kind of, while the code behaves the same as using v:register, it
doesn't actually use v:register due to some clash with vim-cutlass).

While this is added primarily to aid Vim popup users, it doesn't break
anything for Neovim floating win or preview window users as the buffer
is not modifiable, and it seems a handy mapping for all users anyway.